added release function

+++ Alex/coloursensor.h
@@ -42,7 +42,7 @@ void readColour(unsigned long &red, unsigned long &green, unsigned long &blue)
 }
 */
 
-void readColour(unsigned float &red, unsigned float &green, unsigned float &blue) {
+void readColour(unsigned long &red, unsigned long &green, unsigned long &blue) {
  pinMode(23, OUTPUT);
   pinMode(24, OUTPUT);
   pinMode(25, OUTPUT);
@@ -61,21 +61,21 @@ void readColour(unsigned float &red, unsigned float &green, unsigned float &blue
   // Read Red Pulse Width
   digitalWrite(25,LOW);
   digitalWrite(26,LOW);
-  red = (double)pulseIn(35,LOW);
+  red = pulseIn(35,LOW);
   // Delay to stabilize sensor
   delay(20);
 
   // Read Green Pulse Width
   digitalWrite(25,HIGH);
   digitalWrite(26,HIGH);
-  green = (double)pulseIn(35,LOW);
+  green = pulseIn(35,LOW);
   // Delay to stabilize sensor
   delay(20);
 
   // Read Blue Pulse Width
   digitalWrite(25,LOW);
   digitalWrite(26,HIGH);
-  blue = (double)pulseIn(35,LOW);
+  blue = pulseIn(35,LOW);
   // Delay to stabilize sensor
   delay(20);
